#6 코드리뷰 반영 - Regex를 common 폴더로 이동 - ServiceImpl에서 builder 패턴 dto로 분리
