+++ src/main/java/com/gregori/controller/mypage/MypageController.java
@@ -19,26 +19,26 @@ import lombok.RequiredArgsConstructor;
 
 @Controller
 @RequiredArgsConstructor
-@RequestMapping("/mypage")
+@RequestMapping("/mypage/member-info")
 public class MypageController {
 
 	private final MemberService memberService;
 
-	@PostMapping("/member-info")
+	@PostMapping
 	public ResponseEntity<String> updateMember(@RequestBody @Valid MemberUpdateDto mypageUpdateDto) {
 		memberService.updateMember(mypageUpdateDto);
 
 		return ResponseEntity.status(HttpStatus.OK).body("회원 수정에 성공했습니다.");
 	}
 
-	@DeleteMapping("/member-info/{memberId}")
+	@DeleteMapping("/{memberId}")
 	public ResponseEntity<String> deactivateMember(@PathVariable Long memberId) {
 		memberService.deactivateMember(memberId);
 
 		return ResponseEntity.status(HttpStatus.OK).body("회원 탈퇴에 성공했습니다.");
 	}
 
-	@GetMapping("/member-info/{memberId}")
+	@GetMapping("/{memberId}")
 	public ResponseEntity<MemberResponseDto> findMemberById(@PathVariable Long memberId) {
 		MemberResponseDto memberResponseDto = memberService.findMemberById(memberId);
 		return ResponseEntity.status(HttpStatus.OK).body(memberResponseDto);
