+++ src/main/java/com/gregori/dto/member/MemberResponseDto.java
@@ -2,10 +2,14 @@ package com.gregori.dto.member;
 
 import java.time.ZonedDateTime;
 
+import com.gregori.domain.member.Member;
+
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
+@NoArgsConstructor
 public class MemberResponseDto {
 	private Long id;
 	private String email;
@@ -20,5 +24,15 @@ public class MemberResponseDto {
 		this.name = name;
 		this.status = status;
 		this.createdAt = createdAt;
+	}
+
+	public MemberResponseDto toEntity(Member member) {
+		return MemberResponseDto.builder()
+			.id(member.getId())
+			.name(member.getName())
+			.email(member.getEmail())
+			.status(member.getStatus().toString())
+			.createdAt(member.getCreatedAt())
+			.build();
 	}
 }
